no-op: Test access to Internal modules. ...
This commit allows the tests to have access to the internal (i.e.
"hidden") modules. The tests _should_ act on observables from the
public interface, so I don't anticipate directly testing anything
exposed from the hidden modules.

The main reason for doing this isn't to write final tests against
private modules, but rather _during development_ it is sometimes
helpful to be able to use some of that stuff while debugging a new
test.