refactor: Fix test schema. ...
The type of `emails.user_id` was wrong because it it did not match the
type of `users.id`. This doesn't show up yet as an error (even though
there is a foreign constraint between the two  fields), but it will do
as soon as we try to join across these two tables.